requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task queue and Redis
celery==5.3.4
redis==5.0.1

# Audio processing
openai-whisper==20231117
pydub==0.25.1
librosa==0.10.1

# AI/ML models
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
sentence-transformers==2.2.2

# Text-to-Speech
gTTS==2.4.0
pyttsx3==2.90

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
